Ability to edit and delete cars.

+++ Data/Models/Car.cs
@@ -2,7 +2,7 @@
 public class Car
 {
     public string Id { get; set; } = Guid.NewGuid().ToString();
-    public int? Displacement { get; set; }
+    public double? Displacement { get; set; }
     public byte[] Picture { get; set; } = default!;
     public string PictureType { get; set; } = default!;
     public Model Model { get; set; } = default!;
